board: artik710_raptor: dfu: Provide correct Vendor and Product ID
This patch is picked up from 8fc171318e35(samsung: dfu: Provide
correct Vendor and Product IDs for UMS gadget).

It is necessary to provide the same Vendor and Product IDs as the one in
the original Linux kernel code.
Without this change the USB mass storage gadget is not working with Windows7.

Change-Id: I6568438f49ff8b265a1694afd251a3d2d021e9e6
Signed-off-by: Chanho Park <chanho61.park@samsung.com>