board: artik530_raptor: dfu: Provide correct Vendor and Product ID
This patch is picked up from 8fc171318e35(samsung: dfu: Provide
correct Vendor and Product IDs for UMS gadget).

It is necessary to provide the same Vendor and Product IDs as the one in
the original Linux kernel code.
Without this change the USB mass storage gadget is not working with Windows7.

Change-Id: Ib5030fde891cb089cd9fbfed789033d0707da05a
Signed-off-by: Chanho Park <chanho61.park@samsung.com>