artik530_raptor: configs: configurable TEXT_BASE
This patch enables TEXT_BASE, SDRAM BASE, SDRAM_SIZE to be set via
the command line.

If use secure boot, the address of text_base will be different.
Therefore, we changed it to offset method so that it can be set
through build option.

    $ make ... CONFIG_SYS_TEXT_BASE=0x91000000 \
        KCPPFLAGS="-DSDRAM_BASE=0x91000000 -DSDRAM_SIZE=0x20000000"

Must apply it with patches that modify environment variables.

Change-Id: I0167b067cf0b3e4b8693a6f682aa5a6d0130d10e
Signed-off-by: Bon-gyu, KOO <freestyle@nexell.co.kr>