arm: s5p6818: set pad functions for debug console
This patch adds to initialize gpio pad functions according to debug
console serial IDs. Actually, the debug console has been initialized
from bl1 stage. However, to change the bl1 binary only for the debug
console seemed to be not necessary. For a developer who want to use
a different debug console IDs, we also initialize the GPIO pad
functions in the u-boot stage.

Change-Id: I61da0f1192d4ce55f135ae68cd56a31869ac5266
Signed-off-by: Chanho Park <chanho61.park@samsung.com>