arm: dts: add adtik711s device tree files
This patch adds artik711s device tree files in order to support
the artik711s raptor board. The main difference from artik710 is
changes of the eMMC and sdcard port number. In order to support it,
we should override dw_mmc0 and dw_mmc2 nodes and switch the address of
the each register and values.

Change-Id: I3cad3271fbe89d57529d94cc165350d71543d169
Signed-off-by: Chanho Park <chanho61.park@samsung.com>