board: nexell: s5p6818: add artik711s_raptor board
This patch adds an artik711s_raptor board which is based on s5p6818
SoC. It has several differences with artik710_raptor board such as
eMMC/SD card port. In order to support the board specific feature of
the artik711s_raptor board, we need to split the artik711s_raptor board
directory.

Change-Id: Iecc94deed11ec348c3eba8c3a5eb012d6f9ec44d
Signed-off-by: Chanho Park <chanho61.park@samsung.com>